Remove obsolete migratestatus tool and move shared code to statusreconciler.
