--- maintenance/migratestatus/.gitignore
@@ -1,1 +1,0 @@
-migratestatus
